Enforce 16KB page size for all ABIs
It seems that ANDROID_SUPPORT_FLEXIBLE_PAGE_SIZES=ON will only use 16KB
for 64-bit ABIs which is likely all we care, however when running
check_elf_alignment.sh over my app's APK it reports there are unaligned
libs (for x86 and armeabi-v7a). Even if not strictly necessary, it would
be nice to see the tool report "ELF Verification Successful" and not
have to look further.